Rename fetchApi folder and improve build
