+++ packages/utils/package.json
@@ -2,9 +2,9 @@
   "name": "@kununu/utils",
   "author": "kununu",
   "license": "ISC",
-  "version": "1.0.1",
+  "version": "1.1.0-beta.1",
   "description": "Hosts react utils to be used cross application.",
-  "main": "dist/",
+  "main": "dist/index.js",
   "scripts": {
     "build": "babel --extensions '.ts,.tsx' src --out-dir dist --ignore 'src/**/*.spec.ts'",
     "clean": "../../scripts/clean.sh",
@@ -37,16 +37,16 @@
     "@types/react": "17.0.5",
     "@typescript-eslint/eslint-plugin": "4.28.4",
     "@typescript-eslint/parser": "4.28.4",
+    "fetch-mock": "9.11.0",
     "jest": "26.6.3",
     "jsdom": "16.5.1",
     "react": "16.13.1",
     "react-dom": "16.13.1",
-    "typescript": "4.3.4",
-    "fetch-mock": "9.11.0"
+    "typescript": "4.3.4"
   },
   "peerDependencies": {
+    "http-status-codes": ">=2.1.4",
     "isomorphic-fetch": ">=3.0.0",
-    "qs": ">=6.10.1",
-    "http-status-codes": ">=2.1.4"
+    "qs": ">=6.10.1"
   }
 }
